requirements.txt
streamlit
streamlit-webrtc
openai
langchain
python-dotenv
soundfile
av
numpy
